--- @types/@ember/polyfills/index.d.ts
@@ -1,5 +1,0 @@
-/**
- * Copy properties from a source object to a target object.
- * https://github.com/DefinitelyTyped/DefinitelyTyped/issues/38681
- */
-export const assign = Object.assign;
