+++ @types/@ember/version.d.ts
@@ -1,1 +1,1 @@
-export const VERSION: string = '';
+export const VERSION: string;
