+++ @types/@glimmer/tracking.d.ts
@@ -1,3 +1,3 @@
-export { tracked } from '@glimmer/tracking';
-
 export function cached(target: object, key: string, desc: PropertyDescriptor): void;
+
+export function tracked(target: object, key: string): void;
