+++ packages/adapter/addon/-private/build-url-mixin.js
@@ -1,0 +1,5 @@
+import Mixin from '@ember/object/mixin';
+
+import mixinProps from './build-url-mixin-base';
+
+export default Mixin.create(mixinProps);
