+++ packages/store/addon/-private/ts-interfaces/record-data-store-wrapper.ts
@@ -4,7 +4,7 @@ import {
   HasManyRelationshipFieldsFor,
   RecordField,
   RecordType,
-  RelatedType,
+  RelatedField,
   RelationshipFieldsFor,
 } from '@ember-data/types/utils';
 
@@ -19,44 +19,90 @@ import type { AttributesSchema, RelationshipsSchema } from './record-data-schema
  * Provides encapsulated API access to a minimal subset of store service's
  * functionality for RecordData implementations.
  *
+ * This wrapper is provided as an argument to the hook [`Store#createRecordDataFor`](/ember-data/release/classes/Store/methods/createRecordDataFor?anchor=createRecordDataFor)
+ * to be used by instantiated RecordData instances for interacting with the store or other
+ * RecordData instances.
+ *
  * @class RecordDataStoreWrapper
  * @public
  */
 export interface RecordDataStoreWrapper<R extends ResolvedRegistry> {
+  /**
+   * @method relationshipsDefinitionFor
+   * @public
+   * @param type
+   */
   relationshipsDefinitionFor<T extends RecordType<R>>(type: T): RelationshipsSchema<R, T>;
+
+  /**
+   * @method attributesDefinitionFor
+   * @public
+   * @param type
+   */
   attributesDefinitionFor<T extends RecordType<R>>(type: T): AttributesSchema<R, T>;
 
   /**
-   * update the `id` for the record of type `modelName` with the corresponding `clientId`
+   * update the `id` for the record of type `modelName` with the corresponding `lid`
    * This operation can only be done for records whose `id` is `null`.
    *
    * @method setRecordId
    * @public
+   * @public
    */
-  setRecordId<T extends RecordType<R>>(type: T, id: string, clientId: string): void;
+  setRecordId<T extends RecordType<R>>(type: T, id: string, lid: string): void;
 
-  disconnectRecord<T extends RecordType<R>>(type: T, id: string | null, clientId: string): void;
-  disconnectRecord<T extends RecordType<R>>(type: T, id: string, clientId?: string | null): void;
-  disconnectRecord<T extends RecordType<R>>(type: T, id: string | null, clientId?: string | null): void;
+  /**
+   * @method disconnectRecord
+   * @public
+   * @param type
+   * @param id
+   * @param lid
+   */
+  disconnectRecord<T extends RecordType<R>>(type: T, id: string | null, lid: string): void;
+  disconnectRecord<T extends RecordType<R>>(type: T, id: string, lid?: string | null): void;
+  disconnectRecord<T extends RecordType<R>>(type: T, id: string | null, lid?: string | null): void;
 
-  isRecordInUse<T extends RecordType<R>>(type: T, id: string | null, clientId: string): boolean;
-  isRecordInUse<T extends RecordType<R>>(type: T, id: string, clientId?: string | null): boolean;
-  isRecordInUse<T extends RecordType<R>>(type: T, id: string | null, clientId?: string | null): boolean;
+  /**
+   * @method isRecordInUse
+   * @public
+   * @param type
+   * @param id
+   * @param lid
+   */
+  isRecordInUse<T extends RecordType<R>>(type: T, id: string | null, lid: string): boolean;
+  isRecordInUse<T extends RecordType<R>>(type: T, id: string, lid?: string | null): boolean;
+  isRecordInUse<T extends RecordType<R>>(type: T, id: string | null, lid?: string | null): boolean;
 
+  /**
+   * @method notifyPropertyChange
+   * @public
+   * @param type
+   * @param id
+   * @param lid
+   * @param key
+   */
   notifyPropertyChange<T extends RecordType<R>, F extends RecordField<R, T> = RecordField<R, T>>(
     type: T,
     id: string | null,
-    clientId: string | null,
+    lid: string | null,
     key: F
   ): void;
 
+  /**
+   * @method notifyHasManyChange
+   * @public
+   * @param type
+   * @param id
+   * @param lid
+   * @param key
+   */
   notifyHasManyChange<
     T extends RecordType<R>,
     F extends HasManyRelationshipFieldsFor<R, T> = HasManyRelationshipFieldsFor<R, T>
   >(
     type: T,
     id: string | null,
-    clientId: string,
+    lid: string,
     key: F
   ): void;
   notifyHasManyChange<
@@ -65,7 +111,7 @@ export interface RecordDataStoreWrapper<R extends ResolvedRegistry> {
   >(
     type: T,
     id: string,
-    clientId: string | null | undefined,
+    lid: string | null | undefined,
     key: F
   ): void;
   notifyHasManyChange<
@@ -74,22 +120,37 @@ export interface RecordDataStoreWrapper<R extends ResolvedRegistry> {
   >(
     type: T,
     id: string | null,
-    clientId: string | null | undefined,
+    lid: string | null | undefined,
     key: F
   ): void;
 
+  /**
+   * @method recordDataFor
+   * @public
+   * @param type
+   * @param id
+   * @param lid
+   */
   recordDataFor<T extends RecordType<R>>(type: T, id: string, lid?: string | null): RecordData<R, T>;
   recordDataFor<T extends RecordType<R>>(type: T, id: string | null, lid: string): RecordData<R, T>;
   recordDataFor<T extends RecordType<R>>(type: T): RecordData<R, T>;
   recordDataFor<T extends RecordType<R>>(type: T, id?: string | null, lid?: string | null): RecordData<R, T>;
 
+  /**
+   * @method notifyBelongsToChange
+   * @public
+   * @param type
+   * @param id
+   * @param lid
+   * @param key
+   */
   notifyBelongsToChange<
     T extends RecordType<R>,
     F extends BelongsToRelationshipFieldsFor<R, T> = BelongsToRelationshipFieldsFor<R, T>
   >(
     type: T,
     id: string | null,
-    clientId: string,
+    lid: string,
     key: F
   ): void;
   notifyBelongsToChange<
@@ -98,7 +159,7 @@ export interface RecordDataStoreWrapper<R extends ResolvedRegistry> {
   >(
     type: T,
     id: string,
-    clientId: string | null | undefined,
+    lid: string | null | undefined,
     key: F
   ): void;
   notifyBelongsToChange<
@@ -107,19 +168,48 @@ export interface RecordDataStoreWrapper<R extends ResolvedRegistry> {
   >(
     type: T,
     id: string | null,
-    clientId: string | null | undefined,
+    lid: string | null | undefined,
     key: F
   ): void;
 
+  /**
+   * @method inverseForRelationship
+   * @public
+   * @param type
+   * @param key
+   */
   inverseForRelationship<T extends RecordType<R>, F extends RelationshipFieldsFor<R, T>>(
     type: T,
     key: F
-  ): RelatedType<R, T, F> | null;
+  ): RelatedField<R, T, F> | null;
 
+  /**
+   * @method inverseIsAsyncForRelationship
+   * @public
+   * @param type
+   * @param key
+   */
   inverseIsAsyncForRelationship<T extends RecordType<R>, F extends RelationshipFieldsFor<R, T>>(
     type: T,
     key: F
   ): boolean;
-  notifyErrorsChange<T extends RecordType<R>>(type: T, id: string | null, clientId: string | null): void;
-  notifyStateChange<T extends RecordType<R>>(type: T, id: string | null, clientId: string | null, key?: string): void;
+
+  /**
+   * @method notifyErrorsChange
+   * @public
+   * @param type
+   * @param id
+   * @param lid
+   */
+  notifyErrorsChange<T extends RecordType<R>>(type: T, id: string | null, lid: string | null): void;
+
+  /**
+   * @method notifyStateChange
+   * @public
+   * @param type
+   * @param id
+   * @param lid
+   * @param key
+   */
+  notifyStateChange<T extends RecordType<R>>(type: T, id: string | null, lid: string | null, key?: string): void;
 }
